Roles

+++ trsaints-frontend-api/Controllers/ProjectController.cs
@@ -1,6 +1,9 @@
 using AutoMapper;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using trsaints_frontend_api.Authorization;
+using trsaints_frontend_api.Context;
 using trsaints_frontend_api.DTOs;
 using trsaints_frontend_api.Entities;
 using trsaints_frontend_api.Repositories.Interfaces;
@@ -11,14 +14,20 @@ namespace trsaints_frontend_api.Controllers;
 [Route("api/[controller]")]
 [Authorize(AuthenticationSchemes = "Bearer")]
 [ProducesResponseType(StatusCodes.Status401Unauthorized)]
-public class ProjectController: ControllerBase
+public class ProjectController: DI_BaseController
 {
     private readonly IProjectRepository _projectRepository;
     private readonly IMapper _mapper;
 
-    public ProjectController(IProjectRepository repository, IMapper mapper)
+    public ProjectController(
+        AppDbContext context,
+        IAuthorizationService authorizationService,
+        UserManager<ApplicationUser> userManager,
+        IProjectRepository projectRepository,
+        IMapper mapper)
+        : base(context, authorizationService, userManager)
     {
-        _projectRepository = repository;
+        _projectRepository = projectRepository;
         _mapper = mapper;
     }
 
@@ -67,9 +76,16 @@ public class ProjectController: ControllerBase
             return BadRequest();
 
         var project = _mapper.Map<Project>(projectDto);
+        var isAuthorized = await AuthorizationService.AuthorizeAsync(
+            User, project,
+            ResourceOperations.Create);
+
+        if (!isAuthorized.Succeeded)
+            return Forbid();
+        
         await _projectRepository.AddAsync(project);
 
-        return Ok(_mapper.Map<ProjectDTO>(project));
+        return Created($"/api/Project/{project.Id}", projectDto);
     }
         
     [HttpPut("{id:int}")]
@@ -80,6 +96,15 @@ public class ProjectController: ControllerBase
     {
         if (id != projectDto.Id || !ModelState.IsValid)
             return BadRequest();
+        
+        var project = _mapper.Map<Project>(projectDto);
+        var isAuthorized = await AuthorizationService.AuthorizeAsync(
+            User, project,
+            ResourceOperations.Update);
+
+        if (!isAuthorized.Succeeded)
+            return Forbid();
+
             
         await _projectRepository.UpdateAsync(_mapper.Map<Project>(projectDto));
 
@@ -93,9 +118,16 @@ public class ProjectController: ControllerBase
     {
         var project = await _projectRepository.GetByIdAsync(id);
 
+        var isAuthorized = await AuthorizationService.AuthorizeAsync(
+            User, project,
+            ResourceOperations.Delete);
+
+        if (!isAuthorized.Succeeded)
+            return Forbid();
+        
         await _projectRepository.RemoveAsync(project.Id);
 
-        return Ok();
+        return NoContent();
     }
 
     [HttpGet]
